backend/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
aiosqlite
pydantic
python-dotenv
python-multipart
langchain
langchain-core
langchain-openai
langchain-pinecone
langchain-chroma
langchain-community
pinecone-client
chromadb
openai
